Add LEGO processing capabilities to titiler
- Introduced new JSON configuration for LEGO services, including global mosaic, tile instructions, and brick quantization.
- Implemented legofication process in `lego.py` to convert images to LEGO-style representations.
- Added functions for generating LEGO instructions and calculating brick quantities from legofied images.
- Integrated coastline shapefile for accurate land/water classification in Europe.
- Enhanced color matching for LEGO bricks using CIEDE2000 color difference.
