--- services/lego.json
@@ -1,106 +1,0 @@
-{
-  "parameters": [],
-  "process_graph": {
-    "1": {
-      "process_id": "load_collection_and_reduce",
-      "result": false,
-      "arguments": {
-        "bands": [
-          "B04",
-          "B03",
-          "B02"
-        ],
-        "id": "sentinel-2-global-mosaics",
-        "spatial_extent": {
-          "east": {
-            "from_parameter": "spatial_extent_east"
-          },
-          "north": {
-            "from_parameter": "spatial_extent_north"
-          },
-          "south": {
-            "from_parameter": "spatial_extent_south"
-          },
-          "west": {
-            "from_parameter": "spatial_extent_west"
-          },
-          "crs": {
-            "from_parameter": "spatial_extent_crs"
-          }
-        },
-        "temporal_extent": [
-          "2024-04-01T00:00:00Z",
-          "2024-04-30T23:59:59Z"
-        ]
-      }
-    },
-    "2": {
-      "process_id": "save_result",
-      "result": true,
-      "arguments": {
-        "data": {
-          "from_node": "lego"
-        },
-        "format": "PNG",
-        "options": {
-          "datatype": "byte"
-        }
-      }
-    },
-    "3": {
-      "process_id": "apply",
-      "result": false,
-      "arguments": {
-        "data": {
-          "from_node": "1"
-        },
-        "process": {
-          "process_graph": {
-            "1": {
-              "arguments": {
-                "inputMax": 10000,
-                "inputMin": 0,
-                "outputMax": 255,
-                "x": {
-                  "from_parameter": "x"
-                }
-              },
-              "process_id": "linear_scale_range"
-            },
-            "2": {
-              "process_id": "trunc",
-              "arguments": {
-                "x": {
-                  "from_node": "1"
-                }
-              },
-              "result": true
-            }
-          }
-        }
-      }
-    },
-    "color": {
-      "process_id": "color_formula",
-      "result": false,
-      "arguments": {
-        "data": {
-          "from_node": "3"
-        },
-        "formula": "Gamma R 2 Gamma G 2.13 Gamma B 2 Sigmoidal RGB 10 0.24 Saturation 1.1"
-      }
-    },
-    "lego": {
-      "process_id": "legofication",
-      "result": false,
-      "arguments": {
-        "data": {
-          "from_node": "color"
-        },
-        "nbbricks": 4,
-        "bricksize": 64,
-        "water_threshold": 0.1
-      }
-    }
-  }
-}
