+++ titiler/openeo/processes/implementations/__init__.py
@@ -7,6 +7,7 @@ from .dem import *  # noqa
 from .image import *  # noqa
 from .indices import *  # noqa
 from .io import *  # noqa
+from .lego import *  # noqa
 from .math import *  # noqa
 from .reduce import *  # noqa
 from .spatial import *  # noqa
